fix(mesh): run relay-mesh agents on sprout-agent, not goose
The "Run on relay mesh" preset handed the agent the global default
runtime (goose) while setting SPROUT_AGENT_PROVIDER + OPENAI_COMPAT_*
env vars. Goose ignores those and falls back to its own configured
provider, so mesh agents never used the local mesh endpoint.

Source the agent command and MCP from the sprout-agent provider catalog
entry, which reads those env vars (crates/sprout-agent/src/config.rs).
This matches the frontend preset contract (agentCommand: sprout-agent,
mcpCommand: sprout-dev-mcp) the JS test fixture already documented.
